Tags and tests.
- Added test tags: multi_working, multi_broken, multi_unknown, and multi_mixed
- Changed the test run in ci to use tag multi_fixed
- Updated the PR template to make the author check that they updated the tags.
- Fixed/tweaked broken tests I encountered.
- Rearranged some more tests.

Files checked in: .github/pull_request_template.md .github/workflows/tracebase-tests.yml DataRepo/example_data/testing_data/tissues/tissues_with_errors.tsv DataRepo/management/commands/load_tissues.py DataRepo/tests/loading/test_loading_compounds.py DataRepo/tests/loading/test_loading_tissues.py DataRepo/tests/models/test_animal.py DataRepo/tests/models/test_hier_cached_model.py DataRepo/tests/models/test_infusate.py DataRepo/tests/models/test_infusate_tracer.py DataRepo/tests/models/test_model_utilities.py DataRepo/tests/models/test_peak_data.py DataRepo/tests/models/test_tracer.py DataRepo/tests/models/test_tracer_label.py DataRepo/tests/test_customtags.py DataRepo/tests/test_formats.py DataRepo/tests/test_infusate_parsing_validation.py DataRepo/tests/test_models.py DataRepo/tests/test_utils.py DataRepo/tests/test_views.py DataRepo/utils/tissues_loader.py DataRepo/tests/models/test_compounds.py
