+++ .github/pull_request_template.md
@@ -20,4 +20,6 @@ attention to.  E.g. There are significant logic changes in function X.
 - [ ] [Linting passes](https://github.com/Princeton-LSI-ResearchComputing/tracebase/blob/main/CONTRIBUTING.md#linting).
 - [ ] [Migrations created & committed *(or no model changes)*](https://github.com/Princeton-LSI-ResearchComputing/tracebase/blob/main/CONTRIBUTING.md#migration-process)
 - [ ] [Tests implemented *(or no code changes)*](https://github.com/Princeton-LSI-ResearchComputing/tracebase/blob/main/CONTRIBUTING.md#test-implementation)
+  - [ ] New test classes/functions tagged with `@tag("multi_working")`
 - [ ] [All tests pass](https://github.com/Princeton-LSI-ResearchComputing/tracebase/blob/main/CONTRIBUTING.md#quality-control)
+  - [ ] Test function tag `@tag("multi_working")` added and test class/function tags `@tag("multi_broken")`, `@tag("multi_unknown")`, or `@tag("multi_mixed")` removed or changed.
