+++ .github/workflows/tracebase-tests.yml
@@ -75,7 +75,7 @@ jobs:
           python manage.py createcachetable \
           --database validation
       - name: Run tests
-        run: python manage.py test
+        run: python manage.py test --tag multi_working
       - name: Load initial tissue records
         run: |
           python manage.py load_study \
