+++ DataRepo/example_data/testing_data/tissues/tissues_with_errors.tsv
@@ -1,40 +1,9 @@
 Name	Description
-brown_adipose_tissue	brown adipose tissue
-brown_adipose_tissue	a different description
-brain	whole brain
-brain	whole brain
-diaphragm	diaphragm muscle
-	empty name
-gastrocnemius	gastrocnemius muscle
-white_adipose_tissue_gonadal	gonadal white adipose tissue
-heart	heart muscle (ventricle)
-kidney	kidney
-liver	liver
-lung	lung
-pancreas	pancreas
-quadricep	quadricep muscle
-serum_plasma_unspecified_location	serum or plasma - only use when the source is unknown
-small_intestine	small intestine
-soleus	soleus
-skeletal_muscle_unspecified_location	unspecified skeletal muscle - only use this if source muscle is unknown
-colon	colon
-white_adipose_tissue_inguinal	inguinal white adipose tissue
-white_adipose_tissue_unspecified_location	unspecified white adipose tissue - only use this if source fat depot is unknown
-spleen	spleen
-serum_plasma_tail	serum or plasma collected from tail snip
-serum_plasma_artery	serum or plasma collected from any artery
-serum_plasma_portal	serum or plasma collected from portal vein
-thymus	thymus
-ear	outer ear, usually collected as a representative skin sample
-skin	unspecified source of skin - only use this if source of skin is unknown
-stomach	stomach
-cecum	cecum
-cecum_contents	cecum contents
-tibialus_anterior	tibialus anterior muscle
-eye	eyeball
-testicle	testicle
-ovary	ovary
-uterus	uterus
-tumor_nonspecific	unspecified tumor - only use this if other information is unknown
-tumor_HCT116	xenograft tumor of HCT116 cells
-lymph_node	lymph node
+brown_adipose_tissue	a description
+brown_adipose_tissue	a different description should cause an error
+brain	same description is no problem
+brain	same description is no problem
+	empty name should cause an error
+no_tab_no_space_no_problem
+space but no description is a problem	
+space with description is ok	description
