+++ DataRepo/tests/models/test_animal.py
@@ -9,6 +9,7 @@ from DataRepo.tests.tracebase_test_case import TracebaseTestCase
 
 @override_settings(CACHES=settings.TEST_CACHES)
 @tag("animal")
+@tag("multi_working")
 class AnimalTests(TracebaseTestCase):
     def setUp(self):
         Animal.objects.create(
