+++ DataRepo/tests/models/test_tracer_label.py
@@ -1,9 +1,12 @@
+from django.test import tag
+
 from DataRepo.models.compound import Compound
 from DataRepo.models.tracer import Tracer
 from DataRepo.models.tracer_label import TracerLabel
 from DataRepo.tests.tracebase_test_case import TracebaseTestCase
 
 
+@tag("multi_working")
 class TracerLabelTests(TracebaseTestCase):
     def setUp(self):
         glu = Compound.objects.create(name="glucose", formula="C6H12O6")
