+++ DataRepo/tests/test_formats.py
@@ -24,6 +24,7 @@ from DataRepo.models.utilities import get_model_by_name
 from DataRepo.tests.tracebase_test_case import TracebaseTestCase
 
 
+@tag("multi_mixed")
 class FormatsTests(TracebaseTestCase):
     maxDiff = None
 
@@ -173,7 +174,7 @@ class FormatsTests(TracebaseTestCase):
             ("msrun__sample__animal__treatment__name", "Treatment"),
         )
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_getSearchFieldChoicesDict(self):
         basv = SearchGroup()
         sfcd = basv.getSearchFieldChoicesDict()
@@ -224,14 +225,14 @@ class FormatsTests(TracebaseTestCase):
         )
         self.assertTupleEqual(sfct_expected, sfct)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_extractFldPaths(self):
         qry = self.getQueryObject()
         paths = extractFldPaths(qry)
         expected_paths = ["msrun__sample__animal__studies"]
         self.assertEqual(expected_paths, paths)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_splitCommon_hascommon(self):
         fld_path = "msrun__sample__animal__studies"
         reroot_path = "msrun__sample__animal__tracer_compound"
@@ -239,7 +240,7 @@ class FormatsTests(TracebaseTestCase):
         self.assertEqual(common_path, "msrun__sample__animal")
         self.assertEqual("studies", remainder)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_splitCommon_nocommon(self):
         fld_path = "msrun__sample__animal__studies"
         reroot_path = "compounds__synonyms"
@@ -247,13 +248,13 @@ class FormatsTests(TracebaseTestCase):
         self.assertEqual(common_path, "")
         self.assertEqual("msrun__sample__animal__studies", remainder)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_splitPathName(self):
         path, name = splitPathName("msrun__sample__animal__treatment__name")
         self.assertEqual(path, "msrun__sample__animal__treatment")
         self.assertEqual("name", name)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_reRootFieldPath(self):
         fld = "msrun__sample__animal__studies__name"
         reroot_instance_name = "CompoundSynonym"
@@ -262,7 +263,7 @@ class FormatsTests(TracebaseTestCase):
         expected_fld = "compound__peak_groups__msrun__sample__animal__studies__name"
         self.assertEqual(expected_fld, rerooted_fld)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_reRootQry(self):
         qry = self.getQueryObject2()
         qry_backup = deepcopy(qry)
@@ -278,13 +279,13 @@ class FormatsTests(TracebaseTestCase):
         self.assertEqual(qry, qry_backup, msg="qry must be left unchanged")
         self.assertEqual(expected_qry, new_qry)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_pathToModelInstanceName(self):
         pgsv = PeakGroupsFormat()
         mi = pgsv.pathToModelInstanceName("msrun__sample__animal__studies")
         self.assertEqual("Study", mi)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_getTrueJoinPrefetchPathsAndQrys(self):
         qry = self.getQueryObject2()
         basv = SearchGroup()
@@ -356,7 +357,7 @@ class FormatsTests(TracebaseTestCase):
 
         self.assertEqual(expected_prefetches, prefetches)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_getFullJoinAnnotations(self):
         basv = SearchGroup()
         fmt = "pgtemplate"
@@ -376,7 +377,7 @@ class FormatsTests(TracebaseTestCase):
         expected_annots = [{annot_name: F("compounds__pk")}]
         self.assertEqual(expected_annots, annots)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_getDistinctFields(self):
         basv = SearchGroup()
         fmt = "pgtemplate"
@@ -400,7 +401,7 @@ class FormatsTests(TracebaseTestCase):
         ]
         self.assertEqual(expected_distincts, distincts)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_getDistinctFields_split_all(self):
         """
         Ensures that meta ordering fields are expanded to real database fields.  I.e. it tests that fields from every
@@ -438,13 +439,13 @@ class FormatsTests(TracebaseTestCase):
         ]
         self.assertEqual(expected_distincts, distincts)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_getFKModelName(self):
         pgf = PeakGroupsFormat()
         mdl_name = pgf.getFKModelName(CompoundSynonym(), "compound")
         self.assertEqual("Compound", mdl_name)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_getOrderByFields_instance(self):
         pgsv = PeakDataFormat()
         mdl_inst = "PeakData"
@@ -464,7 +465,7 @@ class FormatsTests(TracebaseTestCase):
         expected_order_bys = ["peak_group__name", "corrected_abundance"]
         self.assertEqual(expected_order_bys, order_bys)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_getOrderByFields_model(self):
         pgsv = PeakGroupsFormat()
         mdl = "Compound"
@@ -473,7 +474,7 @@ class FormatsTests(TracebaseTestCase):
         expected_order_bys = ["name"]
         self.assertEqual(expected_order_bys, order_bys)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_getOrderByFields_both(self):
         pgsv = PeakGroupsFormat()
         mdl_inst = "MeasuredCompound"
@@ -484,7 +485,7 @@ class FormatsTests(TracebaseTestCase):
         ):
             pgsv.getOrderByFields(mdl_inst_nm=mdl_inst, model_name=mdl)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_getOrderByFields_neither(self):
         pgsv = PeakGroupsFormat()
         mdl_inst = "MeasuredCompound"
@@ -495,7 +496,7 @@ class FormatsTests(TracebaseTestCase):
         ):
             pgsv.getOrderByFields(mdl_inst_nm=mdl_inst, model_name=mdl)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_createFilterGroup(self):
         got = createFilterGroup()
         expected = {
@@ -506,7 +507,7 @@ class FormatsTests(TracebaseTestCase):
         }
         self.assertEqual(expected, got)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_createFilterCondition(self):
         fld = "fldtest"
         ncmp = "ncmptest"
@@ -522,7 +523,7 @@ class FormatsTests(TracebaseTestCase):
         }
         self.assertEqual(expected, got)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_appendFilterToGroup(self):
         fld = "fldtest"
         ncmp = "ncmptest"
@@ -547,7 +548,7 @@ class FormatsTests(TracebaseTestCase):
         }
         self.assertEqual(expected, got)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_getStatsParams(self):
         pgsv = PeakGroupsFormat()
         stats = pgsv.getStatsParams()
@@ -559,7 +560,7 @@ class FormatsTests(TracebaseTestCase):
         }
         self.assertEqual(expected_i2, got)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_getAllBrowseData(self):
         """
         Test that test_getAllBrowseData returns all data for the selected format.
@@ -615,7 +616,7 @@ class FormatsTests(TracebaseTestCase):
         }
         return tval, qry
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_createNewBasicQuery(self):
         """
         Test createNewBasicQuery creates a correct qry
@@ -631,7 +632,7 @@ class FormatsTests(TracebaseTestCase):
         self.maxDiff = None
         self.assertEqual(newqry, qry)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_searchFieldToDisplayField(self):
         """
         Test that searchFieldToDisplayField converts Study.id to Study.name
@@ -822,7 +823,7 @@ class FormatsTests(TracebaseTestCase):
             "show": True,
         }
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_performQueryStats(self):
         """
         Test that performQuery returns a correct stats structure
@@ -833,7 +834,7 @@ class FormatsTests(TracebaseTestCase):
         expected_stats = self.getExpectedStats()
         self.assertEqual(expected_stats, stats)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_getQueryStats(self):
         """
         Test that getQueryStats returns a correct stats structure
@@ -848,7 +849,7 @@ class FormatsTests(TracebaseTestCase):
         expected = full_stats["data"]
         self.assertEqual(expected, got)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_constructAdvancedQuery(self):
         """
         Test that constructAdvancedQuery returns a correct Q expression
@@ -858,7 +859,7 @@ class FormatsTests(TracebaseTestCase):
         expected_q = Q(msrun__sample__tissue__name__iexact="Brain")
         self.assertEqual(expected_q, q_exp)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_performQuery(self):
         """
         Test that performQuery returns a correct queryset
@@ -885,7 +886,7 @@ class FormatsTests(TracebaseTestCase):
         }
         self.assertEqual(expected_stats, stats)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_performQuery_distinct(self):
         """
         Test that performQuery returns no duplicate root table records when M:M tables queried with multiple matches.
@@ -902,7 +903,7 @@ class FormatsTests(TracebaseTestCase):
         self.assertTrue(cnt < qs.count())
         self.assertEqual(cnt, 1)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_isQryObjValid(self):
         """
         Test that isQryObjValid correctly validates a qry object.
@@ -915,7 +916,7 @@ class FormatsTests(TracebaseTestCase):
         isvalid = isQryObjValid(qry, basv_metadata.getFormatNames().keys())
         self.assertEqual(isvalid, False)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_isValidQryObjPopulated(self):
         """
         Test that isValidQryObjPopulated correctly interprets the population of a subgroup.
@@ -927,7 +928,7 @@ class FormatsTests(TracebaseTestCase):
         isvalid = isValidQryObjPopulated(qry)
         self.assertEqual(isvalid, False)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_getJoinedRecFieldValue(self):
         """
         Test that getJoinedRecFieldValue gets a value from a joined table
@@ -941,7 +942,7 @@ class FormatsTests(TracebaseTestCase):
         val = basv_metadata.getJoinedRecFieldValue(recs, fmt, mdl, fld, fld, "Fasted")
         self.assertEqual(val, "Fasted")
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_cv_getSearchFieldChoices(self):
         """
         Test getSearchFieldChoices
@@ -978,7 +979,7 @@ class FormatsTests(TracebaseTestCase):
         )
         self.assertEqual(res, choices)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_cv_getKeyPathList(self):
         """
         Test getKeyPathList
@@ -990,7 +991,7 @@ class FormatsTests(TracebaseTestCase):
         kpl = ["msrun", "sample", "animal"]
         self.assertEqual(res, kpl)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_cv_getPrefetches(self):
         """
         Test getPrefetches (which should not return the infusatetracer through model and return paths in order of
@@ -1010,7 +1011,7 @@ class FormatsTests(TracebaseTestCase):
         ]
         self.assertEqual(pfl, res)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_cv_getModelInstances(self):
         """
         Test getModelInstances
@@ -1034,7 +1035,7 @@ class FormatsTests(TracebaseTestCase):
         ]
         self.assertEqual(res, ml)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_cv_getSearchFields(self):
         """
         Test getSearchFields
@@ -1055,7 +1056,7 @@ class FormatsTests(TracebaseTestCase):
         }
         self.assertEqual(sfd, res)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_cv_getDisplayFields(self):
         """
         Test getDisplayFields
@@ -1078,7 +1079,7 @@ class FormatsTests(TracebaseTestCase):
         }
         self.assertEqual(dfd, res)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_cv_getFormatNames(self):
         """
         Test getFormatNames
@@ -1092,7 +1093,7 @@ class FormatsTests(TracebaseTestCase):
         }
         self.assertEqual(fnd, res)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_cv_formatNameOrKeyToKey(self):
         """
         Test formatNameOrKeyToKey
@@ -1102,7 +1103,7 @@ class FormatsTests(TracebaseTestCase):
         res = basv_metadata.formatNameOrKeyToKey(fmt)
         self.assertEqual(res, "pgtemplate")
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_pathStepToPosGroupType_inner_node(self):
         """
         Convert "0-all" to [0, "all"]
@@ -1112,7 +1113,7 @@ class FormatsTests(TracebaseTestCase):
         self.assertEqual(gtype, "all")
         self.assertTrue(not static)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_pathStepToPosGroupType_leaf_node(self):
         """
         Convert "0" to [0, None]
@@ -1122,7 +1123,7 @@ class FormatsTests(TracebaseTestCase):
         self.assertEqual(gtype, None)
         self.assertEqual(static, False)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_rootToFormatInfo_selected(self):
         """
         Convert "pgtemplate-PeakGroups-selected" to ["pgtemplate", "PeakGroups", True]
@@ -1132,7 +1133,7 @@ class FormatsTests(TracebaseTestCase):
         self.assertEqual(name, "PeakGroups")
         self.assertEqual(sel, True)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_rootToFormatInfo_unselected(self):
         """
         Convert "pdtemplate-PeakData" to ["pdtemplate", "PeakData", False]
@@ -1144,7 +1145,7 @@ class FormatsTests(TracebaseTestCase):
 
 
 @tag("search_choices")
-@tag("multi_fixed")
+@tag("multi_working")
 class SearchFieldChoicesTests(TracebaseTestCase):
     def test_get_all_comparison_choices(self):
         base_search_view = Format()
