+++ DataRepo/tests/test_models.py
@@ -11,7 +11,6 @@ from django.test import override_settings, tag
 from DataRepo.models import (
     Animal,
     Compound,
-    CompoundSynonym,
     ElementLabel,
     Infusate,
     InfusateTracer,
@@ -41,6 +40,7 @@ from DataRepo.utils import (
 )
 
 
+@tag("multi_unknown")
 class ExampleDataConsumer:
     def get_sample_test_dataframe(self):
 
@@ -85,123 +85,7 @@ class ExampleDataConsumer:
 
 
 @override_settings(CACHES=settings.TEST_CACHES)
-class CompoundTests(TracebaseTestCase):
-    def setUp(self):
-        Compound.objects.create(
-            name="alanine", formula="C3H7NO2", hmdb_id="HMDB0000161"
-        )
-
-    def test_compound_name(self):
-        """Compound lookup by name"""
-        alanine = Compound.objects.get(name="alanine")
-        self.assertEqual(alanine.name, "alanine")
-
-    def test_compound_hmdb_url(self):
-        """Compound hmdb url"""
-        alanine = Compound.objects.get(name="alanine")
-        self.assertEqual(alanine.hmdb_url, f"{Compound.HMDB_CPD_URL}/{alanine.hmdb_id}")
-
-    def test_compound_atom_count(self):
-        """Compound atom_count"""
-        alanine = Compound.objects.get(name="alanine")
-        self.assertEqual(alanine.atom_count("C"), 3)
-
-    def test_compound_atom_count_zero(self):
-        """Compound atom_count"""
-        alanine = Compound.objects.get(name="alanine")
-        self.assertEqual(alanine.atom_count("F"), 0)
-
-    def test_compound_atom_count_invalid(self):
-        """Compound atom_count"""
-        alanine = Compound.objects.get(name="alanine")
-        with self.assertWarns(UserWarning):
-            self.assertEqual(alanine.atom_count("Abc"), None)
-
-
-@override_settings(CACHES=settings.TEST_CACHES)
-class CompoundSynonymTests(TracebaseTestCase):
-    def setUp(self):
-        self.PRIMARY_COMPOUND = Compound.objects.create(
-            name="hexadecanoic acid", formula="C16H32O2", hmdb_id="HMDB0000220"
-        )
-        # just the act of creating a compound (above) creates two synonyms for
-        # it, in this case
-        self.ALIASES_SETUP_COUNT = 2
-        aliases = ["palmitic acid", "C16:0"]
-        self.ALIASES_SETUP_COUNT += len(aliases)
-        self.PRIMARY_ALIASES = aliases
-        # make synonyms
-        for alias in aliases:
-            CompoundSynonym.objects.create(name=alias, compound=self.PRIMARY_COMPOUND)
-
-        self.SECONDARY_COMPOUND = Compound.objects.create(
-            name="alanine", formula="C3H7NO2", hmdb_id="HMDB0000161"
-        )
-
-    def test_compound_synonym_insertion1(self):
-        #  validates all the aliases created during setUp
-        self.assertTrue(
-            self.PRIMARY_COMPOUND.synonyms.filter(name="hexadecanoic acid").exists()
-        )
-        self.assertTrue(
-            self.PRIMARY_COMPOUND.synonyms.filter(name="Hexadecanoic acid").exists()
-        )
-        for alias in self.PRIMARY_ALIASES:
-            self.assertTrue(self.PRIMARY_COMPOUND.synonyms.filter(name=alias).exists())
-        # setup insertions count
-        self.assertEqual(
-            len(self.PRIMARY_COMPOUND.synonyms.all()), self.ALIASES_SETUP_COUNT
-        )
-
-    def test_compound_synonym_insertion2(self):
-        # test CompoundSynonym's intrinsic class creation method
-        alt_name = "Palmitate"
-        CompoundSynonym.objects.create(name=alt_name, compound=self.PRIMARY_COMPOUND)
-        self.assertTrue(self.PRIMARY_COMPOUND.synonyms.filter(name=alt_name).exists())
-
-    def test_compound_synonym_insertion3(self):
-        # test Compound's utility instance creation method
-        alt_name = "Hexadecanoate"
-        self.PRIMARY_COMPOUND.get_or_create_synonym(alt_name)
-        self.assertTrue(self.PRIMARY_COMPOUND.synonyms.filter(name=alt_name).exists())
-
-    def test_compound_synonym_duplication1(self):
-        # test that duplicate insertion fails
-        with self.assertRaises(IntegrityError):
-            CompoundSynonym.objects.create(
-                name=self.PRIMARY_ALIASES[0], compound=self.PRIMARY_COMPOUND
-            )
-
-    def test_compound_synonym_duplication2(self):
-        # test that attempting to use the same synonym for multiple compounds fails
-        with self.assertRaises(IntegrityError):
-            CompoundSynonym.objects.create(
-                name=self.PRIMARY_ALIASES[0], compound=self.SECONDARY_COMPOUND
-            )
-
-    def test_compound_deletion(self):
-        # compound deletion should remove all synonyms
-        c = Compound.objects.create(
-            name="1-Methylhistidine", formula="C7H11N3O2", hmdb_id="HMDB0000001"
-        )
-        alias = "1 methylhistidine"
-        CompoundSynonym.objects.create(name=alias, compound=c)
-        self.assertTrue(CompoundSynonym.objects.filter(name=alias).exists())
-        c.delete()
-        self.assertFalse(CompoundSynonym.objects.filter(name=alias).exists())
-
-    def test_compound_synonym_deletion(self):
-        # synonym deletion does not alter the compound record
-        c = Compound.objects.create(
-            name="1-Methylhistidine", formula="C7H11N3O2", hmdb_id="HMDB0000001"
-        )
-        alias = "1 methylhistidine"
-        cs = CompoundSynonym.objects.create(name=alias, compound=c)
-        cs.delete()
-        self.assertTrue(Compound.objects.filter(name="1-Methylhistidine").exists())
-
-
-@override_settings(CACHES=settings.TEST_CACHES)
+@tag("multi_broken")
 class StudyTests(TracebaseTestCase, ExampleDataConsumer):
     def setUp(self):
         # Get test data
@@ -369,6 +253,7 @@ class StudyTests(TracebaseTestCase, ExampleDataConsumer):
 
 @override_settings(CACHES=settings.TEST_CACHES)
 @tag("protocol")
+@tag("multi_unknown")
 class ProtocolTests(TracebaseTestCase):
     def setUp(self):
         self.p1 = Protocol.objects.create(
@@ -419,25 +304,7 @@ class ProtocolTests(TracebaseTestCase):
 
 
 @override_settings(CACHES=settings.TEST_CACHES)
-@tag("compound_loading")
-class CompoundValidationLoadingTests(TracebaseTestCase):
-    @classmethod
-    def setUpTestData(cls):
-        call_command("load_study", "DataRepo/example_data/tissues/loading.yaml")
-        call_command(
-            "load_compounds",
-            compounds="DataRepo/example_data/consolidated_tracebase_compound_list.tsv",
-            validate_only=True,
-            verbosity=0,
-        )
-        # validate only; nothing gets loaded
-        cls.ALL_COMPOUNDS_COUNT = 0
-
-    def test_compounds_loaded(self):
-        self.assertEqual(Compound.objects.all().count(), self.ALL_COMPOUNDS_COUNT)
-
-
-@override_settings(CACHES=settings.TEST_CACHES)
+@tag("multi_working")
 class DataLoadingTests(TracebaseTestCase):
     @classmethod
     def setUpTestData(cls):
@@ -845,6 +712,7 @@ class DataLoadingTests(TracebaseTestCase):
 
 
 @override_settings(CACHES=settings.TEST_CACHES)
+@tag("multi_mixed")
 class PropertyTests(TracebaseTestCase):
     @classmethod
     def setUpTestData(cls):
@@ -964,7 +832,7 @@ class PropertyTests(TracebaseTestCase):
         self.assertEqual(peakdata.last().id, peakdata2.last().id)
 
     @tag("fcirc", "serum")
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_missing_serum_sample_peak_data(self):
         animal = self.MAIN_SERUM_ANIMAL
         final_serum_sample = animal.final_serum_sample
@@ -1024,7 +892,7 @@ class PropertyTests(TracebaseTestCase):
                 refeshed_animal.final_serum_tracer_rate_appearance_average_per_animal
             )
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_peak_group_peak_data_1(self):
         peak_group = (
             PeakGroup.objects.filter(compounds__name="glucose")
@@ -1045,7 +913,7 @@ class PropertyTests(TracebaseTestCase):
         self.assertEqual(list(peak_group.normalized_labelings.keys()), ["C"])
         self.assertAlmostEqual(peak_group.normalized_labelings["C"], 0.009119978074)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_peak_group_peak_data_4(self):
         # null original data
         peak_group = (
@@ -1071,7 +939,7 @@ class PropertyTests(TracebaseTestCase):
         self.assertEqual(list(peak_group.normalized_labelings.keys()), ["C"])
         self.assertAlmostEqual(peak_group.normalized_labelings["C"], 0.009119978074)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_peak_group_peak_data_serum(self):
         peak_group = (
             PeakGroup.objects.filter(compounds__name="lysine")
@@ -1091,7 +959,7 @@ class PropertyTests(TracebaseTestCase):
         self.assertEqual(list(peak_group.normalized_labelings.keys()), ["C"])
         self.assertAlmostEqual(peak_group.normalized_labelings["C"], 1)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_no_peak_labeled_elements(self):
         # This creates an animal with a notrogen-labeled tracer (among others)
         call_command(
@@ -1139,7 +1007,7 @@ class PropertyTests(TracebaseTestCase):
         ):
             pg.enrichment_fractions
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_enrichment_fraction_missing_compounds(self):
         peak_group = (
             PeakGroup.objects.filter(compounds__name="lysine")
@@ -1150,7 +1018,7 @@ class PropertyTests(TracebaseTestCase):
         with self.assertWarns(UserWarning):
             self.assertIsNone(peak_group.enrichment_fractions)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_enrichment_fraction_missing_labeled_element(self):
         peak_group = (
             PeakGroup.objects.filter(compounds__name="lysine")
@@ -1166,7 +1034,7 @@ class PropertyTests(TracebaseTestCase):
         with self.assertWarns(UserWarning):
             self.assertIsNone(peak_group.enrichment_fractions)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_peak_group_peak_labeled_elements(self):
         peak_group = (
             PeakGroup.objects.filter(compounds__name="lysine")
@@ -1176,7 +1044,7 @@ class PropertyTests(TracebaseTestCase):
 
         self.assertEqual(["C"], peak_group.peak_labeled_elements)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_peak_group_tracer_labeled_elements(self):
         peak_group = (
             PeakGroup.objects.filter(compounds__name="lysine")
@@ -1186,7 +1054,7 @@ class PropertyTests(TracebaseTestCase):
 
         self.assertEqual(["C"], peak_group.msrun.sample.animal.tracer_labeled_elements)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_normalized_labeling_latest_serum(self):
         peak_group = (
             PeakGroup.objects.filter(compounds__name="glucose")
@@ -1253,7 +1121,7 @@ class PropertyTests(TracebaseTestCase):
         second_peak_data.save()
         self.assertAlmostEqual(peak_group.normalized_labelings["C"], 3.455355083)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_normalized_labeling_missing_serum_peak_group(self):
         peak_group = (
             PeakGroup.objects.filter(compounds__name="glucose")
@@ -1272,7 +1140,7 @@ class PropertyTests(TracebaseTestCase):
         with self.assertWarns(UserWarning):
             self.assertIsNone(peak_group.normalized_labelings)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_normalized_labeling_missing_serum_sample(self):
         peak_group = (
             PeakGroup.objects.filter(compounds__name="glucose")
@@ -1289,7 +1157,7 @@ class PropertyTests(TracebaseTestCase):
         with self.assertWarns(UserWarning):
             self.assertIsNone(peak_group.normalized_labelings)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_peak_data_fraction(self):
         peak_data = (
             PeakGroup.objects.filter(compounds__name="glucose")
@@ -1301,7 +1169,7 @@ class PropertyTests(TracebaseTestCase):
         )
         self.assertAlmostEqual(peak_data.fraction, 0.9952169753)
 
-    @tag("multi_fixed")
+    @tag("multi_working")
     def test_peak_group_total_abundance_zero(self):
         # Test various calculations do not raise exceptions when total_abundance is zero
         peak_group = (
@@ -1457,17 +1325,20 @@ class PropertyTests(TracebaseTestCase):
         # need to invalidate the computed/cached enrichment_fractions, somehow
         pg = self.MAIN_SERUM_ANIMAL.final_serum_sample_tracer_peak_group
         set_cache(pg, "enrichment_fractions", None)
-        # When/if we remove the hier_cached_model strategy and restore the cached_property strategy, we can uncomment:
-        #     # simplest way?
-        #     pg.enrichment_fractions = None
         with self.assertWarns(UserWarning):
             self.assertFalse(pg.can_compute_average_tracer_rates)
 
 
 @override_settings(CACHES=settings.TEST_CACHES)
+@tag("multi_working")
 class MultiTracerLabelPropertyTests(TracebaseTestCase):
     @classmethod
     def setUpTestData(cls):
+        call_command("load_study", "DataRepo/example_data/tissues/loading.yaml")
+        call_command(
+            "load_compounds",
+            compounds="DataRepo/example_data/consolidated_tracebase_compound_list.tsv",
+        )
         call_command(
             "load_animals_and_samples",
             animal_and_sample_table_filename=(
@@ -1493,51 +1364,63 @@ class MultiTracerLabelPropertyTests(TracebaseTestCase):
         self.assertEqual(expected, output)
 
     def test_serum_tracers_enrichment_fractions(self):
-        anml = Animal.objects.get(name="xzl1")
+        anml = Animal.objects.get(name="xzl5")
         output = anml.serum_tracers_enrichment_fractions
-        # TODO: Fix the expected from the test failure output
         expected = {
-            "C": 1,
-            "N": 1,
+            "C": 0.2235244143081364,
+            "N": 0.30075567022988536,
         }
-        self.assertEqual(expected, output)
+        self.assertEqual(len(expected.keys()), len(output.keys()))
+        self.assertAlmostEqual(expected["C"], output["C"])
+        self.assertAlmostEqual(expected["N"], output["N"])
 
     def test_peak_labeled_elements_one(self):
         # succinate has no nitrogen
-        pg = PeakGroup.objects.get(name="succinate")
+        pg = PeakGroup.objects.filter(msrun__sample__name="xzl5_panc").get(
+            name="succinate"
+        )
         output = pg.peak_labeled_elements
         # One common element
         expected = ["C"]
         self.assertEqual(expected, output)
 
     def test_peak_labeled_elements_two(self):
-        pg = PeakGroup.objects.get(name="serine")
+        pg = PeakGroup.objects.filter(msrun__sample__name="xzl5_panc").get(
+            name="glutamine"
+        )
         output = pg.peak_labeled_elements
         expected = ["C", "N"]
         self.assertEqual(expected, output)
 
     def test_enrichment_abundances(self):
-        pg = PeakGroup.objects.get(name="serine")
+        pg = PeakGroup.objects.filter(msrun__sample__name="xzl5_panc").get(
+            name="glutamine"
+        )
         output = pg.enrichment_abundances
-        # TODO: Fix the expected from the test failure output
         expected = {
-            "C": 1,
-            "N": 1,
+            "C": 1369911.2746615328,
+            "N": 6571127.3714690255,
         }
-        self.assertEqual(expected, output)
+        self.assertEqual(len(expected.keys()), len(output.keys()))
+        self.assertAlmostEqual(expected["C"], output["C"])
+        self.assertAlmostEqual(expected["N"], output["N"])
 
     def test_normalized_labelings(self):
-        pg = PeakGroup.objects.get(name="serine")
-        output = pg.enrichment_abundances
-        # TODO: Fix the expected from the test failure output
+        pg = PeakGroup.objects.filter(msrun__sample__name="xzl5_panc").get(
+            name="glutamine"
+        )
+        output = pg.normalized_labelings
         expected = {
-            "C": 1,
-            "N": 1,
+            "C": 0.06287501342027346,
+            "N": 0.2241489339907528,
         }
-        self.assertEqual(expected, output)
+        self.assertEqual(len(expected.keys()), len(output.keys()))
+        self.assertAlmostEqual(expected["C"], output["C"])
+        self.assertAlmostEqual(expected["N"], output["N"])
 
 
 @override_settings(CACHES=settings.TEST_CACHES)
+@tag("multi_broken")
 class TracerRateTests(TracebaseTestCase):
     @classmethod
     def setUpTestData(cls):
@@ -1700,6 +1583,7 @@ class TracerRateTests(TracebaseTestCase):
 
 
 @override_settings(CACHES=settings.TEST_CACHES)
+@tag("multi_working")
 class AnimalAndSampleLoadingTests(TracebaseTestCase):
     @classmethod
     def setUpTestData(cls):
@@ -1735,6 +1619,7 @@ class AnimalAndSampleLoadingTests(TracebaseTestCase):
 
 
 @override_settings(CACHES=settings.TEST_CACHES)
+@tag("multi_working")
 class AccuCorDataLoadingTests(TracebaseTestCase):
     @classmethod
     def setUpTestData(cls):
@@ -1821,6 +1706,7 @@ class AccuCorDataLoadingTests(TracebaseTestCase):
 
 
 @override_settings(CACHES=settings.TEST_CACHES)
+@tag("multi_working")
 class IsoCorrDataLoadingTests(TracebaseTestCase):
     @classmethod
     def setUpTestData(cls):
@@ -2239,10 +2125,15 @@ class IsoCorrDataLoadingTests(TracebaseTestCase):
             isocorr_format=True,
         )
         pg = (
-            PeakGroup.objects.filter(name__exact="succinate")
-            .filter(peak_group_set__filename="alafasted_cor.xlsx")
-            .distinct(["id", "peak_data__labels__element"])
+            PeakGroup.objects.filter(msrun__sample__name="xzl5_panc")
+            .filter(name__exact="serine")
+            .filter(
+                peak_group_set__filename="alafasted_cor.xlsx",
+            )
+            .order_by("id", "peak_data__labels__element")
+            .distinct("id", "peak_data__labels__element")
         )
+
         self.assertEqual(pg.count(), 2)
         self.assertEqual(pg.filter(peak_data__labels__element__exact="C").count(), 1)
         self.assertEqual(pg.filter(peak_data__labels__element__exact="N").count(), 1)
@@ -2250,6 +2141,7 @@ class IsoCorrDataLoadingTests(TracebaseTestCase):
 
 @override_settings(CACHES=settings.TEST_CACHES)
 @tag("load_study")
+@tag("multi_broken")
 class StudyLoadingTests(TracebaseTestCase):
     @classmethod
     def setUpTestData(cls):
@@ -2321,6 +2213,7 @@ class StudyLoadingTests(TracebaseTestCase):
 
 
 @override_settings(CACHES=settings.TEST_CACHES)
+@tag("multi_working")
 class ParseIsotopeLabelTests(TracebaseTestCase):
     @classmethod
     def setUpTestData(cls):
@@ -2428,6 +2321,7 @@ class ParseIsotopeLabelTests(TracebaseTestCase):
 @override_settings(CACHES=settings.TEST_CACHES)
 @tag("animal")
 @tag("loading")
+@tag("multi_unknown")
 class AnimalLoadingTests(TracebaseTestCase):
     """Tests parsing various Animal attributes"""
